Fixing top-face algorithm
I was looking at doing an object-oriented approach and realized I was
rotating all squares on a face with one another. This is incorrect, the
corners should move into corners and the edges into edges. While we're
fixing this, further break down the problem of a face rotation, hoping
to show some pattern in the rotation